requirements.txt
qdrant-client>=1.6.0
sentence-transformers>=2.2.2
langchain>=0.0.335
langchain-community>=0.0.1
langchain-core>=0.1.0
langchain-google-genai>=0.0.5
google-generativeai>=0.3.0
torch>=2.0.0
transformers>=4.30.0
python-dotenv>=1.0.0
accelerate>=0.27.0
bitsandbytes>=0.42.0 